PipeWire capture: Clean latest_available_frame on capture stop
Once we stop capturing, we clear the queue with frames, but we still
keep pointer to the latest available frame. This latest available frame
should be invalidated, otherwise it will be now pointing to non-existent
frame. This is not a problem when we completely stop capturing, but if
we are just restarting it to pick a different source, we can crash on it
if we are asked to provide a frame, but we didn't get any yet from the
new source.

(cherry picked from commit 576ec5a272658bead30d36a992fa579c9b36f62c)

Bug: chromium:400656070
Change-Id: I780392999ebba917aa72dde8a2c9b77bfc21cd0f
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/379920
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Jan Grulich <grulja@gmail.com>
Cr-Original-Commit-Position: refs/heads/main@{#44084}
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/382621
Cr-Commit-Position: refs/branch-heads/7049@{#2}
Cr-Branched-From: caae8ec03bedd938b99c7a11b6455878c4f93e75-refs/heads/main@{#44027}